Rename utility file
