+++ deploy/scripts/src/common/utils/request/static.ts
